+++ lib/src/ropc_client_for_native.dart
@@ -22,7 +22,7 @@ import 'package:simple_jwt_manager/src/static_fields/f_json_keys_to_server.dart'
 class ROPCClientForNative {
   // static parameters
   static const String className = "ROPCClientForNative";
-  static const int version = 2;
+  static const int version = 3;
 
   // parameters
   late final String _registerUrl;
@@ -35,6 +35,7 @@ class ROPCClientForNative {
   late final Duration _connectionTimeout;
   late final Duration _responseTimeout;
   late final int refreshMarginMs;
+  final void Function(Map<String, dynamic> savedData)? updateJwtCallback;
 
   // tokens
   String? _accessToken;
@@ -65,6 +66,12 @@ class ROPCClientForNative {
   /// early by the margin you set here.
   /// By default, if the access token is due to expire within 30 seconds,
   /// it will be automatically refreshed using a refresh token.
+  /// * [updateJwtCallback] : This is a callback that is called whenever
+  /// a managed JWT is updated. It is called after a JWT is retrieved, replaced,
+  /// or deleted.
+  /// The result of the toDict call on this class is passed as an argument to
+  /// the function, and can be saved to compose the savedData used on app restart
+  /// that will be passed on app restart.
   ROPCClientForNative(
       {required String registerURL,
       required String signInURL,
@@ -76,7 +83,8 @@ class ROPCClientForNative {
       bool Function(X509Certificate cert, String host, int port)?
           badCertificateCallback,
       Map<String, dynamic>? savedData,
-      this.refreshMarginMs = 30 * 1000}) {
+      this.refreshMarginMs = 30 * 1000,
+      this.updateJwtCallback}) {
     _registerUrl = UtilCheckURL.validateHttpsUrl(registerURL);
     _signInUrl = UtilCheckURL.validateHttpsUrl(signInURL);
     _refreshUrl = UtilCheckURL.validateHttpsUrl(refreshURL);
@@ -182,6 +190,8 @@ class ROPCClientForNative {
   /// 戻り値にトークンが含まれない場合はアクセス成功フラグが返され、トークンは更新されません。
   ///
   /// * [email] : User mail address. It is used as an ID(User name).
+  /// According to the OAuth2.0 specification,
+  /// this will be sent as the username.
   /// * [password] : pw.
   /// * [scope] : Application specific access permissions passed
   /// in space separated format, e.g. read write, user:follow, etc.
@@ -239,6 +249,8 @@ class ROPCClientForNative {
   /// 本ペッケージの実装ではdeleteUserURLに対してJSONで情報がPOSTされます。
   ///
   /// * [email] : User mail address. It is used as an ID(User name).
+  /// According to the OAuth2.0 specification,
+  /// this will be sent as the username.
   /// * [password] : pw.
   /// * [option] : Other optional parameters.
   Future<ServerResponse> deleteUser(String email, String password,
@@ -289,6 +301,10 @@ class ROPCClientForNative {
     if (tokens.containsKey(FJsonKeysFromServer.refreshToken)) {
       _refreshToken = tokens[FJsonKeysFromServer.refreshToken];
     }
+    // コールバックがあれば起動する
+    if (updateJwtCallback != null) {
+      updateJwtCallback!(toDict());
+    }
   }
 
   /// (en) Obtain a token using
@@ -298,6 +314,8 @@ class ROPCClientForNative {
   /// トークンを取得します。
   ///
   /// * [email] : User mail address. It is used as an ID(User name).
+  /// According to the OAuth2.0 specification,
+  /// this will be sent as the username.
   /// * [password] : pw.
   /// * [scope] : Application specific access permissions passed
   /// in space separated format, e.g. read write, user:follow, etc.
@@ -413,6 +431,10 @@ class ROPCClientForNative {
           _scope = null;
           _tokenType = null;
         }
+        // コールバックがあれば起動する
+        if (updateJwtCallback != null) {
+          updateJwtCallback!(toDict());
+        }
         return r;
       case EnumServerResponseStatus.timeout:
       case EnumServerResponseStatus.serverError:
@@ -509,5 +531,9 @@ class ROPCClientForNative {
     _scope = null;
     _tokenType = null;
     _refreshToken = null;
+    // コールバックがあれば起動する
+    if (updateJwtCallback != null) {
+      updateJwtCallback!(toDict());
+    }
   }
 }